requirements.txt
# Telegram Bot Framework
aiogram==3.4.1
aiohttp==3.9.1
httpx==0.26.0  # For YandexART API calls

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0  # PostgreSQL async driver
alembic==1.13.1  # Database migrations
pgvector==0.2.4  # Vector search extension

# AI APIs
openai==1.10.0
google-generativeai==0.3.2
anthropic==0.18.1
PyJWT==2.8.0  # For YandexGPT authentication
cryptography==42.0.0  # Required by PyJWT for RS256

# Vector Database & Embeddings
chromadb==0.4.22
sentence-transformers==2.3.1

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0
telethon==1.34.0  # For Telegram scraping

# Scheduling
apscheduler==3.10.4

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Image Processing
Pillow==10.2.0  # For image compositing (watermarks, logos)

# Configuration
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
PyYAML==6.0.1  # For frontmatter parsing in knowledge base

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Logging
loguru==0.7.2

# Web Framework (Partner Panel)
fastapi==0.109.0
uvicorn==0.27.0

# Testing (optional)
pytest==8.0.0
pytest-asyncio==0.23.4
aiosqlite==0.19.0  # SQLite async driver for testing
matplotlib==3.8.2  # For HTML reports with charts
Jinja2==3.1.3  # For HTML templating
